wip: github: Migrate from Travis CI to Github Actions
On travis-ci.com we are now charged for each minute.  We only got 10000
credits in a trial plan, which we used up with a few builds.  Minutes also
cost a different amount of credits on different platforms: 10 on Linux,
but 50 on macOS (installing the dependencies on macOS alone took 12-15
minutes on Travis for some reason, works much faster on Github's runners).

The 'sw_collector' test case had to be disabled because we can't access
/usr/local/share on the Github build hosts (the process is just blocked
in readdir() and eventually times out).

wip: not sure if the sonarcloud test can be migrated easily, their
recommendation is to use travis for C builds.

LGTM might be easier, there exists an example workflow.

caching via ccache etc. is missing yet.

the android build will get its separate workflow

we won't be able to test on different architectures.